Collapse high correlation SNPs

Collapse high correlation SNPs

collapseHighCorSNPs(
  sum_stat,
  ld_mat,
  ld_mat2 = NULL,
  thresh = 0.95,
  score = NULL,
  plot = TRUE
)

Arguments

sum_stat
list of summary statistic tables

ld_mat
list of LD matrices

ld_mat2
optional second list of LD matrices (for different populations). it will be returned alongside the first ld_mat, which is used for the collapsing. the second list of LD matrices is just subset to the same set of SNPs as the first

score
name of a column of sum_stat data.frames with a score, such that collapsing will choose the highest score SNP per collapsed cluster. Otherwise, if set to NULL, the first SNP will be used

plot
logical, draw a before/after grid of plots

threshold
threshold on absolute value of correlation for collapsing, e.g. will collapse if SNPs are more correlated (or anti-correlated) than this amount

Value

list with modified ld_mat and sum_stat lists (and ld_mat2 if provided)